+++ tests/testthat/test-factorial.R
@@ -15,7 +15,7 @@ test_that("Factorial", {
   my_inquiry <- declare_inquiry(interaction = mean(Y_Z_T4 - Y_Z_T3) - mean(Y_Z_T2 - Y_Z_T1))
 
   my_estimator <- declare_estimator(Y ~ Z1 + Z2 + Z1 * Z2,
-    model = lm_robust,
+    .method = lm_robust,
     term = "Z1:Z2"
   )
 
@@ -42,6 +42,6 @@ test_that("Factorial", {
 
   expect_equal(diagnosis %>% get_simulations %>% dim, c(2, 14))
 
-  expect_equal(diagnosis %>%  get_diagnosands %>% dim, c(1, 12))
+  expect_equal(diagnosis %>%  get_diagnosands %>% dim, c(1, 13))
 
 })